feat(mirta-rollup): add ESNext config for NPM
