+++ packages/mirta-rollup/package.json
@@ -39,12 +39,15 @@
     "@babel/preset-env": "^7.28.3",
     "@dotenv-run/rollup": "^1.3.7",
     "@mirta/polyfills": "workspace:*",
+    "@rollup/plugin-commonjs": "^28.0.6",
     "@rollup/plugin-babel": "^6.0.4",
     "@rollup/plugin-multi-entry": "^6.0.1",
     "@rollup/plugin-node-resolve": "^16.0.1",
+    "@rollup/plugin-typescript": "^12.1.4",
     "babel-plugin-array-includes": "^2.0.3",
     "del": "^8.0.1",
     "magic-string": "^0.30.19",
+    "rollup-plugin-dts": "^6.2.3",
     "rollup-plugin-typescript2": "^0.36.0",
     "typescript": "^5.8.3"
   },
